Only provide a single index to `each` callback (d3.alt).
Hopefully, this should avoid confusion as to what the various indexes mean.
Note that since the callback is guaranteed to be invoked in-order, it's
possible with a little work to infer the group index and the global index. But
this at least gets people in the habit of only depending on the local index.