Allow env-js to be require'd.
This fixes a bug with the usage of `require` in env-js, such that env-js now
uses proper relative paths rather than unshifting "." onto `require.paths`. The
behavior of `require` is documented here:

  <http://nodejs.org/docs/v0.4.2/api/modules.html#all_Together...>

With this change in place, it's now possible to use env-js simply by require'ing
it from a script, rather than having env-js host and eval.

Also, this commit changes the initialization of env-js, such that the initial
document is immediately available when the env-js module is loaded. And, the
event loop is disabled so that env-js shuts down cleanly. This probably breaks
some parts of env-js, but it makes it much easier to use so I'm going to play
with it a bit.