Don't postpone immediate timers.
Previously, rescheduling an immediate timer would cause it to be postponed again
by a minimum of 24ms. This is undesirable if the timer is being scheduled in
response to user events, such as mousemoves, that can occur very frequently. Now
we schedule an early callback once, and don't touch it if it's rescheduled. In
addition, the timer loop is now smarter about switching between short and long
callbacks. If there's a short delay (< 24ms) until the next timer fires, the
timer will use requestAnimationFrame. For longer delays, setTimeout is used.