Improve stability of charge forces.
Previously, we were modifying the node's x & y position as we were computing
the charge forces. Unfortunately, this causes drift because those positions are
subsequently used to compute other forces. Now we modify the node's previous
position in the opposite direction, which has the same ultimate effect but
improves the stability of the calculation.

This commit also optimizes the force layout such that gravity and charge forces
are not calculated if the corresponding constants are zero.