Minor optimisation and Chrome bug workaround.
There appears to be a bug in Chrome causing the whole tab to crash when
setting svgNode.className.baseVal = "" for a newly-created element
without a non-empty class name:

  https://code.google.com/p/chromium/issues/detail?id=138552

This fix is a minor optimisation: if the current class is empty then
removing a class should be a no-op.  It also conveniently works around
the Chrome bug.