Only attempt to resample if distance < δ.
This improves performance in the case where our input data is already of
a high enough resolution, even when subject to distortion. It’s unlikely
that the Douglas–Peucker estimate (perpendicular distance) will be
greater than δ if the two input points are already closer than δ. We
could probably even increase the threshold a little more e.g. 2 * δ
seems reasonable.