Restore undefined behavior for d3.bisect.
Rather than return NaN, this change restores the previous undefined behavior
when an unorderable value is used for bisection. After discussing with Jason, I
think this is more consistent given that d3.bisect also requires that the input
array is sortable, and therefore only contains orderable values. Thus, it is the
responsibility of the caller to check that the search value is defined before
doing bisection, and so this commit fixes the threshold scale rather than the
bisect method. (Also, this is consistent with the existing quantile scale.)