Simpler logic for arc padding.
Rather than dynamically increasing the inner radius for thin wedges, preserve
the inner radius but relax the requirement for sides parallel with the adjacent
arc. As long as the inner radius is reasonably big, the behavior is pleasing
and consistent.

The recommended minimum inner radius when using padding is outerRadius *
padAngle / sin(θ), where θ is the angle of the smallest arc (without padding).
For example, if the outerRadius is 200 pixels and the padAngle is 0.02 radians,
a reasonable θ is 0.04 radians and a reasonable innerRadius is 100 pixels.